+++ static/js/admin/src/ResultObjectActionMenu.js
@@ -189,7 +189,7 @@ define('ResultObjectActionMenu', [ 'jquery', 'jquery-ui', 'StringUtilities',  'A
 		// Export actions
 		if (!isContentRoot) {
 			items["sepexport"] = "";
-			if (metadata.type !== 'File' && $.inArray('viewHidden', metadata.permissions) != -1 && metadata.counts.child <= 50000) {
+			if (metadata.type !== 'File' && $.inArray('viewHidden', metadata.permissions) != -1) {
 				items["exportCSV"] = {name : 'Export as CSV'};
 			}
